requirements.txt
fastapi
uvicorn[standard]
python-multipart
firebase-admin
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
beautifulsoup4
